+++ orangecontrib/text/corpus.py
@@ -320,8 +320,6 @@ class Corpus(Table):
             else:
                 var = ContinuousVariable(f, compute_value=cv)
             var.sparse = sparse     # don't pass this to constructor so this works with Orange < 3.8.0
-            if cv is not None:      # set original variable for cv
-                cv.variable = var
             if isinstance(var_attrs, dict):
                 var.attributes.update(var_attrs)
             additional_attributes.append(var)
@@ -642,30 +640,6 @@ class Corpus(Table):
         else:  # orig is not Corpus
             new._set_unique_titles()
             new._infer_text_features()
-
-    def __eq__(self, other):
-        warnings.warn(
-            "Corpus's __eq__  is deprecated and will be removed in Orange3-text"
-            "1.11. Equality operator will return True only for the same objects.",
-            FutureWarning
-        )
-        def arrays_equal(a, b):
-            if sp.issparse(a) != sp.issparse(b):
-                return False
-            elif sp.issparse(a) and sp.issparse(b):
-                return (a != b).nnz == 0
-            else:
-                return np.array_equal(a, b)
-
-        return (self.text_features == other.text_features and
-                self._dictionary == other._dictionary and
-                np.array_equal(self._tokens, other._tokens) and
-                arrays_equal(self.X, other.X) and
-                arrays_equal(self.Y, other.Y) and
-                arrays_equal(self.metas, other.metas) and
-                np.array_equal(self.pos_tags, other.pos_tags) and
-                self.domain == other.domain and
-                self.ngram_range == other.ngram_range)
 
 
 if summarize: