+++ orangecontrib/text/tests/test_vectorization_base.py
@@ -1,10 +1,14 @@
+import os
 import unittest
 
 import numpy as np
 import scipy.sparse as sp
 
 from orangecontrib.text.corpus import Corpus
-from orangecontrib.text.vectorization.base import BaseVectorizer
+from orangecontrib.text.vectorization.base import (
+    BaseVectorizer,
+    VectorizationComputeValue,
+)
 
 
 class BaseVectorizationTest(unittest.TestCase):
@@ -32,3 +36,33 @@ class BaseVectorizationTest(unittest.TestCase):
 
         for a in c2.domain.attributes:
             self.assertIn('foo', a.attributes)
+
+
+class TestVectorizationComputeValue(unittest.TestCase):
+    def test_unpickling_old_pickle(self):
+        """
+        Before orange3-text version 1.12.0 variable was wrongly set to current
+        variable (variable that has this compute value attached) instead of
+        original variable which caused fails after latest changes in core
+        Orange. Since variable from VectorizationComputeValue is never used in
+        practice we do not set it anymore (it is always None for
+        VectorizationComputeValue).
+        Anyway it is still set in pickles create before 1.12.0. With this test
+        we test that old pickle with variables that have VectorizationComputeValue
+        are un-pickled correctly.
+        """
+        path = os.path.join(
+            os.path.dirname(os.path.abspath(__file__)), "data", "old-bow-pickle.pkl"
+        )
+        data = Corpus.from_file(path)
+        self.assertEqual(len(data), 3)
+        self.assertIsInstance(data.domain["!"].compute_value, VectorizationComputeValue)
+        self.assertIsInstance(
+            data.domain["aboard"].compute_value, VectorizationComputeValue
+        )
+        self.assertIsNone(data.domain["!"].compute_value.variable)
+        self.assertIsNone(data.domain["aboard"].compute_value.variable)
+
+
+if __name__ == "__main__":
+    unittest.main()